Moved fake_embeddings, added more comments to search and changed entrypoint for embeddings
