+++ pkg/vmcp/optimizer/store.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/mark3labs/mcp-go/server"
 
+	"github.com/stacklok/toolhive/pkg/vmcp/optimizer/internal/similarity"
 	sqlitestore "github.com/stacklok/toolhive/pkg/vmcp/optimizer/internal/sqlite_store"
 	"github.com/stacklok/toolhive/pkg/vmcp/optimizer/internal/types"
 )
@@ -92,9 +93,22 @@ func (s *InMemoryToolStore) Search(_ context.Context, query string, allowedTools
 	return matches, nil
 }
 
+// SQLiteStoreConfig configures the SQLite-backed ToolStore.
+// When nil is passed to NewSQLiteToolStore, only FTS5 search is used.
+type SQLiteStoreConfig struct {
+	// EmbeddingDimension enables semantic search with deterministic fake
+	// embeddings of the given dimensionality. Zero disables semantic search.
+	EmbeddingDimension int
+}
+
 // NewSQLiteToolStore creates a new ToolStore backed by SQLite for search.
 // The store uses an in-memory SQLite database with shared cache for concurrent access.
-// If embeddingClient is non-nil, semantic search is enabled alongside FTS5.
-func NewSQLiteToolStore(embeddingClient EmbeddingClient) (ToolStore, error) {
-	return sqlitestore.NewSQLiteToolStore(embeddingClient)
+// If cfg is nil or EmbeddingDimension is zero, only FTS5 search is used.
+// Otherwise, semantic search is enabled alongside FTS5 using the configured embedding dimension.
+func NewSQLiteToolStore(cfg *SQLiteStoreConfig) (ToolStore, error) {
+	var embClient types.EmbeddingClient
+	if cfg != nil && cfg.EmbeddingDimension > 0 {
+		embClient = similarity.NewFakeEmbeddingClient(cfg.EmbeddingDimension)
+	}
+	return sqlitestore.NewSQLiteToolStore(embClient)
 }
